+++ web/src/filter.ts
@@ -295,12 +295,12 @@ export class Filter {
     _sorted_term_types?: string[] = undefined;
     _predicate?: (message: Message) => boolean;
     _can_mark_messages_read?: boolean;
+    requires_adjustment_for_moved_with_target?: boolean;
 
     constructor(terms: NarrowTerm[]) {
-        this._terms = this.fix_terms(terms);
-        if (this.has_operator("channel")) {
-            this._sub = stream_data.get_sub_by_name(this.operands("channel")[0]!);
-        }
+        this._terms = terms;
+        this.setup_filter(terms);
+        this.requires_adjustment_for_moved_with_target = this.has_operator("with");
     }
 
     static canonicalize_operator(operator: string): string {
@@ -387,6 +387,38 @@ export class Filter {
             operator,
             operand,
         };
+    }
+
+    static ensure_channel_topic_terms(orig_terms: NarrowTerm[]): NarrowTerm[] {
+        // In presence of `with` term without channel or topic terms in the narrow, the
+        // narrow is populated with the channel and toipic terms through this operation,
+        // so that `with` can be used as a standalone operator to target conversation.
+        const with_term = orig_terms.find((term: NarrowTerm) => term.operator === "with");
+
+        if (!with_term) {
+            return orig_terms;
+        }
+
+        const updated_terms = [...orig_terms];
+        let channel_term = updated_terms.find(
+            (term: NarrowTerm) => Filter.canonicalize_operator(term.operator) === "channel",
+        );
+        let topic_term = updated_terms.find(
+            (term: NarrowTerm) => Filter.canonicalize_operator(term.operator) === "topic",
+        );
+
+        if (!topic_term) {
+            topic_term = {operator: "topic", operand: ""};
+            const with_index = updated_terms.indexOf(with_term);
+            updated_terms.splice(with_index, 0, topic_term);
+        }
+
+        if (!channel_term) {
+            channel_term = {operator: "channel", operand: ""};
+            const topic_index = updated_terms.indexOf(topic_term);
+            updated_terms.splice(topic_index, 0, channel_term);
+        }
+        return updated_terms;
     }
 
     /* We use a variant of URI encoding which looks reasonably
@@ -576,6 +608,7 @@ export class Filter {
             "channels-public",
             "channel",
             "topic",
+            "with",
             "dm",
             "dm-including",
             "sender",
@@ -775,6 +808,11 @@ export class Filter {
 
     static adjusted_terms_if_moved(raw_terms: NarrowTerm[], message: Message): NarrowTerm[] | null {
         if (message.type !== "stream") {
+            // BUG: We should be replacing the channel/topic/dm
+            // operators with the singular dm operator for the
+            // conversation in question. This isn't very important,
+            // because direct messages cannot be moved, but it should
+            // be fixed, since it is triggerable incorrect behavior.
             return null;
         }
 
@@ -783,6 +821,8 @@ export class Filter {
 
         const adjusted_terms = [];
         let terms_changed = false;
+
+        raw_terms = Filter.ensure_channel_topic_terms(raw_terms);
 
         for (const term of raw_terms) {
             const adjusted_term = {...term};
@@ -810,6 +850,13 @@ export class Filter {
         }
 
         return adjusted_terms;
+    }
+
+    setup_filter(terms: NarrowTerm[]): void {
+        this._terms = this.fix_terms(terms);
+        if (this.has_operator("channel")) {
+            this._sub = stream_data.get_sub_by_name(this.operands("channel")[0]!);
+        }
     }
 
     equals(filter: Filter, excluded_operators?: string[]): boolean {
@@ -929,6 +976,7 @@ export class Filter {
             "channels-web-public",
             "not-channels-web-public",
             "near",
+            "with",
         ]);
 
         for (const term of term_types) {
@@ -963,6 +1011,10 @@ export class Filter {
         // it is limited by the user's message history. Therefore, we check "channel"
         // and "topic" together to ensure that the current filter will return all the
         // messages of a conversation.
+        if (_.isEqual(term_types, ["channel", "topic", "with"])) {
+            return true;
+        }
+
         if (_.isEqual(term_types, ["channel", "topic"])) {
             return true;
         }
@@ -1180,6 +1232,7 @@ export class Filter {
         const term_types = this.sorted_term_types();
         if (
             (term_types.length === 3 && _.isEqual(term_types, ["channel", "topic", "near"])) ||
+            (term_types.length === 3 && _.isEqual(term_types, ["channel", "topic", "with"])) ||
             (term_types.length === 2 && _.isEqual(term_types, ["channel", "topic"])) ||
             (term_types.length === 1 && _.isEqual(term_types, ["channel"]))
         ) {
@@ -1362,15 +1415,27 @@ export class Filter {
     fix_terms(terms: NarrowTerm[]): NarrowTerm[] {
         terms = this._canonicalize_terms(terms);
         terms = this._fix_redundant_is_private(terms);
+        terms = this._fix_redundant_with_dm(terms);
         return terms;
     }
 
     _fix_redundant_is_private(terms: NarrowTerm[]): NarrowTerm[] {
+        // Every DM is a DM, so drop `is:dm` if on a DM conversation.
         if (!terms.some((term) => Filter.term_type(term) === "dm")) {
             return terms;
         }
 
         return terms.filter((term) => Filter.term_type(term) !== "is-dm");
+    }
+
+    _fix_redundant_with_dm(terms: NarrowTerm[]): NarrowTerm[] {
+        // Because DMs can't move, the `with` operator is a noop on a
+        // DM conversation.
+        if (terms.some((term) => Filter.term_type(term) === "dm")) {
+            return terms.filter((term) => Filter.term_type(term) !== "with");
+        }
+
+        return terms;
     }
 
     _canonicalize_terms(terms_mixed_case: NarrowTerm[]): NarrowTerm[] {
@@ -1511,5 +1576,28 @@ export class Filter {
             // not narrowed to starred messages
             !this.has_operand("is", "starred")
         );
+    }
+
+    try_adjusting_for_moved_with_target(message?: Message): void {
+        // If we have the message named in a `with` operator
+        // available, either via parameter or message_store,
+        if (!this.requires_adjustment_for_moved_with_target) {
+            return;
+        }
+
+        if (!message) {
+            const message_id = Number.parseInt(this.operands("with")[0]!, 10);
+            message = message_store.get(message_id);
+        }
+
+        if (!message) {
+            return;
+        }
+
+        const adjusted_terms = Filter.adjusted_terms_if_moved(this._terms, message);
+        if (adjusted_terms) {
+            this.setup_filter(adjusted_terms);
+        }
+        this.requires_adjustment_for_moved_with_target = false;
     }
 }